requirements.txt
discord.py[voice]
google-genai
PyNaCl
python-dotenv